Fixed some user registration flow styles after rebase
