+++ src/components/v5/common/CreateUserForm/CreateUserForm.tsx
@@ -1,14 +1,27 @@
 import React from 'react';
 import { useFormContext } from 'react-hook-form';
+import { defineMessages } from 'react-intl';
 
 import Button from '~v5/shared/Button';
 import { Input } from '~v5/common/Fields';
 import { formatText } from '~utils/intl';
 
-import CreateUserFormHeader from './CreateUserFormHeader';
 import { MAX_USERNAME_LENGTH } from './validation';
+import { HeaderRow } from '../CreateColonyWizard/shared';
 
 const displayName = 'common.CreateUserForm';
+
+const MSG = defineMessages({
+  heading: {
+    id: `${displayName}.heading`,
+    defaultMessage: 'Create your Colony profile',
+  },
+  description: {
+    id: `${displayName}.description`,
+    defaultMessage:
+      'Connecting an email address enhances your Colony experience, such as receiving notifications about activity, mentions, and comments.',
+  },
+});
 
 const CreateUserForm = () => {
   const {
@@ -22,7 +35,7 @@ const CreateUserForm = () => {
 
   return (
     <>
-      <CreateUserFormHeader />
+      <HeaderRow heading={MSG.heading} description={MSG.description} />
       <Input
         name="emailAddress"
         register={register}
@@ -54,9 +67,9 @@ const CreateUserForm = () => {
       <Button
         text={{ id: 'button.continue' }}
         type="submit"
-        mode="solidBlack"
+        mode="primarySolid"
         disabled={!isValid || isSubmitting}
-        className="mt-12"
+        className="mt-3"
       />
     </>
   );
